Fix redis health check ping.
Previous to this PR, we witnessed a failed redis ping but after redis
recovered, the health check /ping never corrected itself.
Incident: https://alerts.newrelic.com/accounts/907948/incidents/6561272

Test Case added: To prove this out, I initially added a test case where
the redis instance came back online and expected the health check to
recover. However it did not.

This is because previously, the /ping endpoint the redis client instance
for the ping was using the 1) default config and was a single Go obj.

The single Go obj did not know how to reconnect after a failed
connection.
Fix: this was fixed by using the same redis pool as the session store.

The default config did not use any time outs so any requests to redis
would wait forever.
Fix: Added timeouts for the read, write and connection attempts.